Reloacted the component files and created the menu for couse management.
