+++ src/components/Navbar.js
@@ -12,6 +12,7 @@ const Navbar = () => {
         <li><Link to="/programs">Programs</Link></li>
         <li><Link to="/profile">Profile</Link></li>
         <li><Link to="/courses">Courses</Link></li>
+        <li><Link to="/coursemanagement">Courses Management</Link></li>
         <li><Link to="/signup">Signup</Link></li>
       </ul>
     </nav>
